DRIVER ALERT SYSTEM
This project helps to alert the driver if they were sleepy while driving the
vehicle. This project capture frame from real-time and detect the face and
eye to determine whether they sleepy or not. If driver is sleepy it will raise a
beep sound.
I have learned different algorithm for detecting face and eyes of
humans
python and opencv have used in the project.
Helpful for me to learn concepts of image processing